Proof of concept: make migrator and central cross compilable
