+++ central/globaldb/v2backuprestore/formats/roxdbv1/bolt.go
@@ -1,3 +1,5 @@
+//go:build amd64
+
 package roxdbv1
 
 import (
